# docs/swagger.yaml
definitions:
...
  models.Device:
    properties:
      city:
        type: string
      date:
        type: string
      deviceName:
        type: string
      ip:
        type: string
      token:
        type: string
    type: object
  models.DeviceListResponse:
    properties:
      current:
        example: current device token
        type: string
      devices:
        items:
          $ref: '#/definitions/models.Device'
        type: array
    type: object
...
  swagger.DevicesResponse:
    properties:
      data:
        $ref: '#/definitions/models.DeviceListResponse'
      msg:
        example: ""
        type: string
      ok:
        example: true
        type: boolean
    type: object
  swagger.EmptyData:
    type: object
...
  swagger.FailedValidationResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.validationErrors'
      msg:
        example: Please review your entered data
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.NotFoundResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: 404 not found!
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.SuccessResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: Successful message
        type: string
      ok:
        example: true
        type: boolean
    type: object
...
  swagger.UnauthenticatedResponse:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyData'
      msg:
        example: You must login first!
        type: string
      ok:
        example: false
        type: boolean
    type: object
...
  swagger.validationErrors:
    properties:
      errors:
        additionalProperties:
          type: string
        example:
          field1: This field is required
          field2: This field must be numeric
        type: object
    type: object
info:
  contact: {}
paths:
...
  /profile/change-password:
    post:
      consumes:
      - application/json
      description: Change Password , authentication required
      parameters:
      - description: Current user password
        in: query
        name: currentPassword
        required: true
        type: string
      - description: 'password that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: password
        required: true
        type: string
      - description: 'repeatPassword that have at least 8 length and contain an alphabet
          and number '
        in: query
        name: repeatPassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: change-password
      tags:
      - profile
  /profile/devices:
    get:
      consumes:
      - application/json
      description: return logged in devices in user's account , authentication required
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.DevicesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
      summary: devices
      tags:
      - profile
  /profile/terminate-device:
    post:
      consumes:
      - application/json
      description: jwt terminate-device , atuhentication required
      parameters:
      - description: token of the device that we want to remove
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.DevicesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: terminate-device
      tags:
      - profile
  /profile/terminate-devices-except-me:
    post:
      consumes:
      - application/json
      description: terminate all devices execpt current device , atuhentication required
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.DevicesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.NotFoundResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: terminate-devices-except-me
      tags:
      - profile
  /profile/upload-profile-picture:
    post:
      consumes:
      - application/json
      description: Upload profile picture , authentication required
      parameters:
      - description: file of image
        in: formData
        name: picture
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthenticatedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.FailedValidationResponse'
      summary: upload-profile-picture
      tags:
      - profile
swagger: "2.0"